pyproject.toml
[project]
name = "elfrfdet"
version = "0.1.0"
description = "Random Forest ELF malware detector — template for lolday (.text[:256] feature)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [{ name = "bolin8017" }]
keywords = ["malware", "detection", "elf", "random-forest", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "islab-malware-detector[mlflow]>=0.5.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "pyelftools>=0.30",
    "structlog>=24.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "ruff>=0.1.0"]

[project.scripts]
elfrfdet = "elfrfdet.cli:app"

[project.urls]
Homepage = "https://github.com/bolin8017/elfrfdet"
Repository = "https://github.com/bolin8017/elfrfdet.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/elfrfdet"]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
